requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
pypdf==3.17.4

# Text processing
nltk==3.8.1
tiktoken==0.5.2

# Embeddings & Vector Store
sentence-transformers==2.2.2
chromadb==0.4.22
faiss-cpu==1.7.4

# LLM Integration
openai==1.6.1
anthropic==0.7.8
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.10

# API & Web
streamlit==1.29.0
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Utilities
numpy==1.26.2
pandas==2.1.4
tqdm==4.66.1
tenacity==8.2.3

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
